+++ src/commands/insert_statement.cpp
@@ -18,7 +18,7 @@ cmd::InsertStatement::InsertStatement() : SqlStatement(INSERT) {}
 st_e::TableRow cmd::InsertStatement::get_row() const {
     auto table_cols = st_e::StorageEngine::get_instance().get_table_by_name(table_name_).get_columns();
     std::vector<std::shared_ptr<st_e::TableCell>> rows;
-    for (int i = 0; i < columns_names_.size(); ++i) {
+    for (size_t i = 0; i < columns_names_.size(); ++i) {
         auto found_col = table_cols.find(columns_names_[i])->second;
         switch (found_col.type)
         {
@@ -63,8 +63,8 @@ bool cmd::InsertStatement::is_valid() const {
     if (columns_names_.size() != columns_vals_.size()){
         return false;
     }
-    for (int i = 0; i < columns_names_.size(); ++i) {
-        for (int j = i + 1; j < columns_names_.size(); ++j) {
+    for (size_t i = 0; i < columns_names_.size(); ++i) {
+        for (size_t j = i + 1; j < columns_names_.size(); ++j) {
             if (columns_names_[i] == columns_names_[j]) {
                 return false;
             }
